+++ nengo_dl/tests/test_simulator.py
@@ -241,7 +241,7 @@ def test_train_recurrent(Simulator, truncation, seed):
         sim.run_steps(n_steps, data={inp: x[:minibatch_size]})
 
     assert np.sqrt(np.mean((sim.data[p] - y[:minibatch_size]) ** 2)) < (
-        0.1 if truncation else 0.05
+        0.1 if truncation else 0.055
     )
 
 
@@ -1005,24 +1005,6 @@ def test_train_state_save(Simulator):
         sim2.run_steps(10)
 
     assert np.allclose(sim.data[p], sim2.data[p])
-
-
-def test_gain_bias(Simulator):
-    N = 17
-    D = 2
-
-    gain = np.random.uniform(low=0.2, high=5, size=N)
-    bias = np.random.uniform(low=0.2, high=1, size=N)
-
-    model = nengo.Network()
-    with model:
-        a = nengo.Ensemble(N, D)
-        a.gain = gain
-        a.bias = bias
-
-    with Simulator(model) as sim:
-        assert np.allclose(gain, sim.data[a].gain)
-        assert np.allclose(bias, sim.data[a].bias)
 
 
 def test_simulation_data(Simulator, seed):